Fix state change getting overwritten in Tabs component
Accordting to documentation, "setState() does not immediately mutate this.state but creates a pending state transition. Accessing this.state after calling this method can potentially return the existing value"

The issue was that calling setState() with a new state derived from this.state in componentWillReceiveProps() over wrote the state change called in onClick(). Instead of calling setState() with a new state object, calling with function.