Implement fake fetch cancellation
Current fetch doesn't have abort feature as there is still a discussion pending about promise cancellation. This commit wraps a promise by return an object that contains the promise itself and a method to cancel the promise. Cancellation will result to promise rejection with an object with isCancelled = true property.